% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAttributeList.R
\name{addAttributeList}
\alias{addAttributeList}
\title{Add attributes using a list of vectors to an igraph object}
\usage{
addAttributeList(g, attr, l)
}
\arguments{
\item{g}{an igraph object}

\item{attr}{the name of the attribute}

\item{l}{the list of vectors}
}
\value{
the modified igraph object
}
\description{
Add attributes using a list of vectors to an igraph object
}
\concept{
paxtoolsr
}
