% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIds.R
\name{convertIds}
\alias{convertIds}
\title{Covert IDs Using org.Hs.eg.db}
\usage{
convertIds(ids, from = "UNIPROT", to = "SYMBOL")
}
\arguments{
\item{ids}{a vector of ids}

\item{from}{the original ID type}

\item{to}{the ID type to convert to}
}
\value{
a vector of converted IDs
}
\description{
Covert IDs Using org.Hs.eg.db
}
\examples{
convertIds("TP53", "SYMBOL", "ENTREZID")

}
\concept{
paxtoolsr
}
