% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSifnxIds.R
\name{convertSifnxIds}
\alias{convertSifnxIds}
\title{Convert IDs in a SIFNX}
\usage{
convertSifnxIds(sifnx, participantType = "ProteinReference",
  idType = "NCBI Gene", mapping = NULL, naRm = TRUE)
}
\arguments{
\item{sifnx}{a SIFNX object (e.g. from the downloadPc2 function)}

\item{participantType}{the type of participant on which the conversion will occur.
Important because not all ID types apply to all entities and otherwise those entities
would be labeled as missing an ID.}

\item{idType}{an ID type for conversion (not used if mapping parameter used)}

\item{mapping}{a two column data.frame with columns mapping$PARTICIPANT (old 
IDs to convert from) and mapping$ID (new IDs to convert to)}

\item{naRm}{remove edges where NA's were introduced due to failed conversions}
}
\value{
a SIFNX list with nodes and edges. Only edges will have converted IDs
}
\description{
Convert IDs in a SIFNX
}
\concept{
paxtoolsr
}

