% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSifToGmt.R
\name{convertSifToGmt}
\alias{convertSifToGmt}
\title{Convert SIF to GMT}
\usage{
convertSifToGmt(sif, name = "gmt", returnSmallMolecules = FALSE)
}
\arguments{
\item{sif}{a data.frame representing a SIF (Simple Interaction Format)}

\item{name}{the name of the gene set}

\item{returnSmallMolecules}{a boolean whether to return genes or small molecules in the gene set}
}
\value{
a list with one entry being a vector
}
\description{
Convert SIF to GMT
}
\concept{
paxtoolsr
}
