% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSifToSpia.R
\name{convertSifToSpia}
\alias{convertSifToSpia}
\title{Convert SIF Interaction Types to SPIA types}
\usage{
convertSifToSpia(edges)
}
\arguments{
\item{edges}{a data.frame of interactions; must have INTERACTION_TYPE column}
}
\value{
the edges data.frame with the converted interaction types
}
\description{
Convert SIF Interaction Types to SPIA types
}
\concept{
paxtoolsr
}

