% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDataFrameWithListOfVectors.R
\name{convertToDataFrameWithListOfVectors}
\alias{convertToDataFrameWithListOfVectors}
\title{Convert Results from readSifnx to data.table}
\usage{
convertToDataFrameWithListOfVectors(lst)
}
\arguments{
\item{lst}{a list returned from readSifnx}
}
\value{
a list entries converted to data.table
}
\description{
Convert Results from readSifnx to data.table
}
\details{
The SIFNX format is an evolving format. Older datasets may not have 
 all the columns this function expects. In these cases, the columns will be 
 added with all NULL entries.
}
\examples{
lst <- readSifnx(system.file("extdata", "test_sifnx_250.txt", package="paxtoolsr"))
newSifnx <- convertToDataFrameWithListOfVectors(lst)
}
\concept{
paxtoolsr
}
