% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDF.R
\name{convertToDF}
\alias{convertToDF}
\title{Convert Results from readSifnx to data.frame}
\usage{
convertToDF(lst)
}
\arguments{
\item{lst}{a list returned from readSifnx}
}
\value{
a list entries converted to data.frame
}
\description{
Convert Results from readSifnx to data.frame
}
\examples{
sifnx <- readSifnx(system.file("extdata", "test_sifnx.txt", package="paxtoolsr"))
class(sifnx$edges)

dfSifnx <- convertToDF(sifnx)
class(dfSifnx$edges)

dtSifnx <- convertToDT(dfSifnx)
class(dtSifnx$edges)

}
\concept{
paxtoolsr
}
