% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertToDT.R
\name{convertToDT}
\alias{convertToDT}
\title{Convert Results from readSifnx to data.table}
\usage{
convertToDT(lst)
}
\arguments{
\item{lst}{a list returned from readSifnx}
}
\value{
a list entries converted to data.table
}
\description{
Convert Results from readSifnx to data.table
}
\concept{
paxtoolsr
}

