% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile}
\alias{downloadFile}
\title{Check Cache and Download File}
\usage{
downloadFile(baseUrl, fileName, cacheEnv = "PAXTOOLSR_CACHE",
  destDir = NULL)
}
\arguments{
\item{baseUrl}{a string, entire download URL except filename}

\item{fileName}{a string, the filename of file to be downloaded}

\item{cacheEnv}{a string, the environment variable that points to the specific cache}

\item{destDir}{a string, the path where a file should be saved}
}
\value{
a boolean TRUE if the file was downloaded or already exists, FALSE otherwise
}
\description{
Check Cache and Download File
}
\details{
Description of file formats: http://www.pathwaycommons.org/pc2/formats
}
\examples{
downloadFile("http://google.com/", fileName="index.html", destDir=tempdir())

}
\seealso{
\code{\link{readSif}, \link{readBiopax}, \link{readSbgn}, \link{readSifnx}, \link{readGmt}}
}
\concept{
paxtoolsr
}

