% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadPc2}
\alias{downloadPc2}
\title{Download Pathway Commons files (uses menu and cache)}
\usage{
downloadPc2(destDir = NULL)
}
\arguments{
\item{destDir}{a string the destination directory for the file to be
  downloaded (Default: NULL). If NULL, then file will be downloaded to cache
  directory ~/.paxtoolsRCache}
}
\description{
Download Pathway Commons files (uses menu and cache)
}
\concept{
paxtoolsr
}

