% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSignedPC.R
\name{downloadSignedPC}
\alias{downloadSignedPC}
\title{Download a SIF file containing only signed interactions}
\usage{
downloadSignedPC(destDir = NULL)
}
\arguments{
\item{destDir}{a string, the destination directory for the file to be 
downloaded (Default: NULL). If NULL, then file will be downloaded to cache
directory file.path(Sys.getenv("HOME"), ".paxtoolsRCache")}
}
\value{
a SIF containing interactions that are considered signed (i.e. 
  interactions causing an increase on decrease in a molecular species)
}
\description{
Download a SIF file containing only signed interactions
}
\examples{
# downloadSignedPC()

}

