% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIds.R
\name{extractIds}
\alias{extractIds}
\title{Extract IDs from an Extended SIF}
\usage{
extractIds(nodes, participantType = "ProteinReference",
  idType = "hgnc symbol")
}
\arguments{
\item{nodes}{extended SIF nodes entries as a data.table (from convertToDT)}

\item{participantType}{a vector of types of participants to search; 
useful to only search protein (ProteinReference) or small molecule (SmallMoleculeReference) related entries.}

\item{idType}{the type of ID to search for; case-insensitive}
}
\value{
a named vector of the first matches for the given ID type
}
\description{
Extract IDs from an Extended SIF
}
\details{
IMPORTANT: Only the first matching ID will be returned. In some cases, multiple IDs will exist.
}
\examples{
tmp <- readSifnx(system.file("extdata", "test_sifnx.txt", package="paxtoolsr"))
results <- extractIds(tmp$nodes)

}
\concept{
paxtoolsr
}

