% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fromPsimi}
\alias{fromPsimi}
\title{Read PSIMI file}
\usage{
fromPsimi(inputFile, outputFile = NULL, bpLevelArg = 3)
}
\arguments{
\item{inputFile}{a string of the name of the input PSIMI file}

\item{outputFile}{a string of the name of the output BioPAX OWL file}

\item{bpLevelArg}{a string representing the BioPAX level for the output file
(default: NULL)}
}
\value{
an XMLInternalDocument representing a BioPAX OWL file
}
\description{
This function reads in a PSIMI file.
}
\details{
The Proteomics Standard Initiative (PSIMI) format is described at
https://code.google.com/p/psimi/wiki/PsimiTabFormat
}
\examples{
outFile <- tempfile()
results <- fromPsimi(system.file("extdata", "10523676-compact.xml", package="paxtoolsr"),
                                 outFile,
                                 "3")
}
\concept{
paxtoolsr
}

