% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCacheFiles.R
\name{getCacheFiles}
\alias{getCacheFiles}
\title{List files in cache directory}
\usage{
getCacheFiles()
}
\value{
a vector of the files in the cache directory
}
\description{
List files in cache directory
}
\examples{
getCacheFiles()

}
\concept{paxtoolsr}
