% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrorMessage.R
\name{getErrorMessage}
\alias{getErrorMessage}
\title{Get Error Message for a Pathway Commons Error}
\usage{
getErrorMessage(code)
}
\arguments{
\item{code}{a three digit numerical error code}
}
\value{
an error message for the code
}
\description{
Get Error Message for a Pathway Commons Error
}
\examples{
results <- getErrorMessage("452")

}
\concept{paxtoolsr}
