% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighbors.R
\name{getNeighbors}
\alias{getNeighbors}
\title{Get the neighbors of a set of IDs in a BioPAX file}
\usage{
getNeighbors(inputFile, outputFile = NULL, idList)
}
\arguments{
\item{inputFile}{a string with the name of the input BioPAX OWL file}

\item{outputFile}{a string with the name of the output BioPAX OWL file}

\item{idList}{a vector of IDs from the BioPAX OWL file}
}
\value{
an XMLInternalDocument representing a BioPAX OWL file
}
\description{
This function retrieves a set of neighbors for a set of IDs in a BioPAX file.
}
\details{
Only entities in the input BioPAX file will be searched for neighbors.
IDs used must be URIs for the entities of interest.
}
\examples{
outFile <- tempfile()
results <- getNeighbors(system.file("extdata", 
  "raf_map_kinase_cascade_reactome.owl", package="paxtoolsr"), 
  outFile, 
  c("HTTP://WWW.REACTOME.ORG/BIOPAX/48887#PROTEIN2360_1_9606", 
    "HTTP://WWW.REACTOME.ORG/BIOPAX/48887#PROTEIN1631_1_9606")) 

}
\concept{
paxtoolsr
}
