% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{getPc}
\alias{getPc}
\title{Get Pathway Commons BioPAX elements}
\usage{
getPc(uri, format = NULL, verbose = FALSE)
}
\arguments{
\item{uri}{a vector that includes valid/existing BioPAX element's URI (RDF
ID; for utility classes that were "normalized", such as entity refereneces
and controlled vocabularies, it is usually a Idntifiers.org URL. Multiple
IDs are allowed per query, for example,
c("http://identifiers.org/uniprot/Q06609",
"http://identifiers.org/uniprot/Q549Z0") See also about MIRIAM and
Identifiers.org in details.}

\item{format}{output format. Valid options can be found using
\code{\link{pcFormats}}}

\item{verbose}{a boolean, display the command used to query Pathway Commons}
}
\value{
a XMLInternalDocument object
}
\description{
This command retrieves full pathway information for a set of elements such as
pathway, interaction or physical entity given the RDF IDs.
}
\details{
Get commands only retrieve the BioPAX elements that are directly
  mapped to the ID. Use the "traverse query to traverse BioPAX graph and
  obtain child/owner elements.

  Information on MIRIAM and Identifiers.org
  \url{http://www.pathwaycommons.org/pc2/#miriam}
}
\examples{
uri <- "http://identifiers.org/uniprot/O14503"
#results <- getPc(uri)

uri <- c("http://identifiers.org/uniprot/O14503", "http://identifiers.org/uniprot/Q9P2X7")
#results <- getPc(uri, verbose=TRUE)
}
\seealso{
\code{\link{pcFormats}}
}
\concept{
paxtoolsr
}

