% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcUrl.R
\name{getPcUrl}
\alias{getPcUrl}
\title{Get base Pathway Commons URL}
\usage{
getPcUrl()
}
\value{
a string with base Pathway Commons URL
}
\description{
Get base Pathway Commons URL
}
\details{
paxtoolsr will support versions Pathway Commons 5 and later. Old 
versions of the webservice will not be not be operational. Users can parse
older BioPAX outputs as an alternative.
}
\examples{
url <- getPcUrl()

}
\concept{paxtoolsr}
