% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShortestPathSif.R
\name{getShortestPathSif}
\alias{getShortestPathSif}
\title{Get the shortest between two IDs (HGNC or CHEBI)}
\usage{
getShortestPathSif(g, idA, idB, mode = c("all", "out", "in"), weights = NA)
}
\arguments{
\item{idA}{HGNC or CHEBI (CHEBI:XXXXX) ID}

\item{idB}{HGNC or CHEBI (CHEBI:XXXXX) ID}

\item{sif}{a SIF network}
}
\value{
a data.frame representing a SIF network
}
\description{
Get the shortest between two IDs (HGNC or CHEBI)
}
\concept{
paxtoolsr
}

