% Generated by roxygen2 (4.0.2): do not edit by hand
\name{graphPc}
\alias{graphPc}
\title{Get Pathway Commons BioPAX elements}
\usage{
graphPc(kind, source, target = NULL, direction = NULL, limit = NULL,
  format = NULL, datasource = NULL, organism = NULL, verbose = FALSE)
}
\arguments{
\item{kind}{graph query. Valid options can be found using pcGraphQueries. See Details for
information on graph queries.}

\item{source}{source object's URI/ID. Multiple source URIs/IDs are allowed
per query, for example c("http://identifiers.org/uniprot/Q06609",
"http://identifiers.org/uniprot/Q549Z0")
See a note about MIRIAM and Identifiers.org in details}

\item{target}{[Required for PATHSFROMTO graph query] target URI/ID. Multiple
target URIs are allowed per query; for example c("http://identifiers.org/uniprot/Q06609",
"http://identifiers.org/uniprot/Q549Z0")
See a note about MIRIAM and Identifiers.org in details}

\item{direction}{[Optional, for NEIGHBORHOOD and COMMONSTREAM algorithms] -
graph search direction. Valid options: \code{\link{pcDirections}}.}

\item{limit}{graph query search distance limit (default: 1).}

\item{format}{output format. Valid options: \code{\link{pcFormats}}}

\item{datasource}{datasource filter (same as for 'search').}

\item{organism}{organism filter (same as for 'search').}

\item{verbose}{a boolean, display the command used to query Pathway Commons}
}
\value{
depending on the the output format a different object may be returned.
  \code{\link{pcFormats}}
}
\description{
This function will retrieve a set of BioPAX elements given a graph query match.
}
\examples{
results <- graphPc(source="http://identifiers.org/uniprot/O14503",
 kind="neighborhood",
 format="EXTENDED_BINARY_SIF")
}
\seealso{
\code{\link{pcFormats}, \link{pcDirections}}
}
\concept{
paxtoolsr
}

