% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/pathwayCommons.R
=======
% Please edit documentation in R/idMapping.R
>>>>>>> devel
\name{idMapping}
\alias{idMapping}
\title{Map IDs to Primary Uniprot or ChEBI IDs}
\usage{
idMapping(ids, verbose = FALSE)
}
\arguments{
\item{ids}{a vector of IDs}

\item{verbose}{a boolean, display the command used to query Pathway Commons}
}
\value{
a list of where each entry is a HGNC symbol provided and the each
  value is a primary UniProt or ChEBI ID.
}
\description{
Unambiguously maps, e.g., HGNC gene symbols, NCBI Gene, RefSeq, ENS*, and
secondary UniProt identifiers to the primary UniProt accessions, or - ChEBI
and PubChem IDs to primary ChEBI. You can mix different standard ID types in
one query.
}
\details{
This is a specific id-mapping (not general-purpose) for reference
  proteins and small molecules; it was first designed for internal use, such
  as to improve BioPAX data integration and allow for graph queries accept
  not only URIs but also standard IDs. The mapping tables were derived
  exclusively from Swiss-Prot (DR fields) and ChEBI data (manually created
  tables and other mapping types and sources can be added in the future
  versions if necessary).
}
\examples{
genes <- c("BRCA2", "TP53")
#results <- idMapping(genes)
}
\concept{
paxtoolsr
}

