% Generated by roxygen2 (4.0.2): do not edit by hand
\name{integrateBiopax}
\alias{integrateBiopax}
\title{Integrate two BioPAX OWL files (DEPRECATED)}
\usage{
integrateBiopax(inputFile1, inputFile2, outputFile = NULL)
}
\arguments{
\item{inputFile1}{a string of the name of the input BioPAX OWL file}

\item{inputFile2}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string of the name of the output integrated BioPAX
OWL file}
}
\value{
an XMLInternalDocument representing a BioPAX OWL file
}
\description{
This function merges two BioPAX OWL files
}
\details{
This method is deprecated. Use mergeBiopax instead.
}
\examples{
outFile <- tempfile()
results <- integrateBiopax(system.file("extdata", "raf_map_kinase_cascade_reactome.owl",
  package="paxtoolsr"),
  system.file("extdata", "circadian_clock.owl", package="paxtoolsr"),
  outFile)
}
\seealso{
\code{\link{mergeBiopax}}
}
\concept{
paxtoolsr
}

