% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSifInIgraph}
\alias{loadSifInIgraph}
\title{Load SIF as igraph Network}
\usage{
loadSifInIgraph(sif)
}
\arguments{
\item{sif}{a binary SIF as a data.frame with three columns:
  "PARTICIPANT_A", "INTERACTION_TYPE", "PARTICIPANT_B"}
}
\value{
a directed igraph network with interaction types
}
\description{
Load SIF as igraph Network
}
\examples{
results <- readSif(system.file("extdata", "test_sif.txt", package="paxtoolsr"))
g <- loadSifInIgraph(results)
}
\concept{
paxtoolsr
}

