% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCytoscape.R
\name{mapAttributes}
\alias{mapAttributes}
\title{Map Attributes from igraph to Cytoscape JSON}
\usage{
mapAttributes(attr.names, all.attr, i)
}
\arguments{
\item{attr.names}{names of attributes}

\item{all.attr}{all attributes}

\item{i}{index}
}
\value{
attributes
}
\description{
Map Attributes from igraph to Cytoscape JSON
}
\note{
From https://github.com/idekerlab/cy-rest-R/blob/17f748426bb5e48ba4075b9d97318ad582b250da/utility/cytoscape_util.R
}
\concept{
paxtoolsr
}
