% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapValues.R
\name{mapValues}
\alias{mapValues}
\title{Map values from One Vector to Another}
\usage{
mapValues(data, oldValue, newValue)
}
\arguments{
\item{data}{a vector of strings where values will be replaced}

\item{oldValue}{a vector that matches values in the data vector}

\item{newValue}{a vector of new values that will replace the old values}
}
\value{
return the vector with the mapped values. If there was no 
  corresponding entry then replace it with an NA.
}
\description{
Map values from One Vector to Another
}
\examples{
data <- c("A", "B", "C", "X", "Y", "Z") 
oldValue <- LETTERS[1:20]
newValue <- letters[1:20]
results <- mapValues(data, oldValue, newValue)

}
\concept{paxtoolsr}
