% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paxtools.R
\name{mergeBiopax}
\alias{mergeBiopax}
\title{Merges two BioPAX OWL files}
\usage{
mergeBiopax(inputFile1, inputFile2, outputFile = NULL)
}
\arguments{
\item{inputFile1}{a string of the name of the input BioPAX OWL file}

\item{inputFile2}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string of the name of the output merged BioPAX
OWL file (Optional)}
}
\value{
an XMLInternalDocument representing a BioPAX OWL file
}
\description{
This function merges two BioPAX OWL files
}
\details{
Only entities that share IDs will be merged. No additional merging
  occurs on cross-references. Merging may result in warning messages caused
  as a result of redundant actions being checked against by the Java library;
  these messages may be ignored.
}
\examples{
outFile <- tempfile()
results <- mergeBiopax(system.file("extdata", "raf_map_kinase_cascade_reactome.owl",
                       package="paxtoolsr"),
                       system.file("extdata", "circadian_clock.owl",
                       package="paxtoolsr"),
                       outFile)
}
\concept{
paxtoolsr
}

