% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/pathwayCommons.R
=======
% Please edit documentation in R/pcFormats.R
>>>>>>> devel
\name{pcFormats}
\alias{pcFormats}
\title{Acceptable Pathway Commons Formats}
\usage{
pcFormats()
}
\value{
acceptable Pathway Commons formats
}
\description{
A simple function to see valid options
}
\details{
See references.
}
\examples{
pcFormats()
}
\references{
Output Formats Description: \url{http://www.pathwaycommons.org/pc2/help/formats.html}
}
\concept{
paxtoolsr
}

