% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{readBiopax}
\alias{readBiopax}
\title{Read BioPAX files as XML documents}
\usage{
readBiopax(inputFile)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
an XMLInternalDocument
}
\description{
Read BioPAX files as XML documents
}
\examples{
results <- readBiopax(system.file("extdata", "biopax3-short-metabolic-pathway.owl", package="paxtoolsr"))
}
\concept{
paxtoolsr
}

