% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCombinedSifnxFile.R
\name{readCombinedSifnxFile}
\alias{readCombinedSifnxFile}
\title{Read a combined file with nodes and edges}
\usage{
readCombinedSifnxFile(inputFile)
}
\arguments{
\item{inputFile}{}
}
\value{
a list of nodes and edges
}
\description{
Read a combined file with nodes and edges
}
\concept{
paxtoolsr
}
