% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPcPathwaysInfo.R
\name{readPcPathwaysInfo}
\alias{readPcPathwaysInfo}
\title{Read in Pathway Commons Pathways Information}
\usage{
readPcPathwaysInfo(inputFile)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
a list with two items: 1) pathwayChildren and 2) pathwayInfo
}
\description{
Read in Pathway Commons Pathways Information
}
\details{
This file is generally found as pathways.txt.gz (e.g. 
http://www.pathwaycommons.org/archives/PC2/current/pathways.txt.gz)
}
\examples{
#results <- readPcPathwaysInfo(INPUTFILE)

}
\concept{
paxtoolsr
}

