% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSif.R
\name{readSif}
\alias{readSif}
\title{Read in a binary SIF file}
\usage{
readSif(inputFile)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
a data.frame with the interactions in the binary SIF format
}
\description{
Read in a binary SIF file
}
\examples{
results <- readSif(system.file("extdata", "test_sif.txt", package="paxtoolsr"))

}
\concept{paxtoolsr}
