% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{readSifnx}
\alias{readSifnx}
\title{Read in a Extended SIF file}
\usage{
readSifnx(inputFile, cols = rep("character", 6))
}
\arguments{
\item{inputFile}{an inputFile}

\item{cols}{a vector of column classes (Default: rep("character", 6)). This
should only be changed for user-customized SIFNX, and it should be set to
'rep("character", N)', where N is the number of columns in the SIFNX.}
}
\value{
a list with nodes and edges entries
}
\description{
Read in a Extended SIF file
}
\details{
SIFNX files from Pathway Commons commonly come a single file that
includes a tab-delimited sections for nodes and another for edges. The
sections are separated by an empty lines. These sections must be split before
they are read.
}
\examples{
results <- readSifnx(system.file("extdata", "test_sifnx.txt", package="paxtoolsr"))
chebiIds <- lapply(results$nodesUniXref, function(x) { x[which(grepl("CHEBI", x))] })
}
\concept{
paxtoolsr
}

