% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSifnxSmall.R
\name{readSifnxSmall}
\alias{readSifnxSmall}
\title{Read small extended SIFs (SIFNX)}
\usage{
readSifnxSmall(inputFile)
}
\arguments{
\item{inputFile}{}
}
\value{
a list of nodes and edges
}
\description{
Read small extended SIFs (SIFNX)
}
\note{
This function is a dependency free version of readSifnx; this is not suitable for very large files
}
\examples{
sifnx <- readSifnxSmall(system.file("extdata", "test_sifnx.txt", package="paxtoolsr"))

}
\concept{
paxtoolsr
}
