% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{searchPc}
\alias{searchPc}
\title{Search Pathway Commons}
\usage{
searchPc(q, page = 0, datasource = NULL, organism = NULL, type = NULL,
  verbose = FALSE)
}
\arguments{
\item{q}{a keyword, name, external identifier, or a Lucene query string.}

\item{page}{an integer giving the search result page number (N>=0, default:
0)}

\item{datasource}{a vector that is a filter by data source (use names or URIs
of pathway data sources or of any existing Provenance object). If multiple
data source values are specified, a union of hits from specified sources is
returned. For example, datasource as c("reactome", "pid") returns hits
associated with Reactome or PID.}

\item{organism}{a vector that is an organism filter. The organism can be
specified either by official name, e.g. "homo sapiens" or by NCBI taxonomy
id, e.g. "9606". Similar to data sources, if multiple organisms are
declared a union of all hits from specified organisms is returned. For
example organism as c("9606", "10016") returns results for both human and
mice. Only humans, "9606" is officially supported.}

\item{type}{BioPAX class filter. See Details.}

\item{verbose}{a boolean, display the command used to query Pathway Commons}
}
\value{
an XMLInternalDocument with results
}
\description{
This command provides a text search using the Lucene query syntax.
}
\details{
Indexed fields were selected based on most common searches. Some of
these fields are direct BioPAX properties, others are composite
relationships. All index fields are (case-sensitive):comment, ecnumber,
keyword, name, pathway, term, xrefdb, xrefid, dataSource, and organism. The
pathway field maps to all participants of pathways that contain the
keyword(s) in any of its text fields. This field is transitive in the sense
that participants of all sub-pathways are also returned. Finally, keyword is
a transitive aggregate field that includes all searchable keywords of that
element and its child elements - e.g. a complex would be returned by a
keyword search if one of its members has a match. Keyword is the default
field type. All searches can also be filtered by data source and organism. It
is also possible to restrict the domain class using the 'type' parameter.
This query can be used standalone or to retrieve starting points for graph
searches. Search strings are case insensitive unless put inside quotes.

BioPAX classes can be found at \url{http://www.pathwaycommons.org/pc2/#biopax_types}
}
\examples{
\dontrun{
results <- searchPc("Q06609")
results <- searchPc("glycolysis", type="Pathway")
}
}
}
\concept{
paxtoolsr
}

