% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{sifInteractionCategories}
\alias{sifInteractionCategories}
\title{Get a list of categories of SIF interactions}
\usage{
sifInteractionCategories()
}
\value{
a list of interactions in categories
}
\description{
Get a list of categories of SIF interactions
}
\details{
Description of interaction types: http://www.pathwaycommons.org/pc2/formats
Categories provided:
  BetweenProteins,
  BetweenProteinsOther (often from high-throughput experiments),
  BetweenProteinSmallMolecule,
  BetweenSmallMolecules
}
\examples{
sifCat <- sifInteractionCategories()
sifCat[["BetweenProteins"]]
}
\concept{
paxtoolsr
}

