% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paxtoolsr.R
\name{skip_on_bioc}
\alias{skip_on_bioc}
\title{Skip a test if on Bioconductor}
\usage{
skip_on_bioc()
}
\value{
A boolean or NULL is returned
}
\description{
Extension on testthat code
}
\concept{
paxtoolsr
}

