% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{splitSifnx}
\alias{splitSifnx}
\title{Split Extended SIF File}
\usage{
splitSifnx(con, verbose = FALSE)
}
\arguments{
\item{con}{an opened file connection}

\item{verbose}{a boolean, display debugging information}
}
\value{
a list with nodes and edges entries
}
\description{
Split an extended SIF file into nodes and edges
}
\details{
SIFNX files from Pathway Commons commonly come a single file that
includes a tab-delimited sections for nodes and another for edges. The
sections are separated by an empty lines. These sections must be split before
they are read.
}
\examples{
con <- file(system.file("extdata", "test_sifnx.txt", package="paxtoolsr"))
results <- splitSifnx(con, verbose=TRUE)
}
\concept{
paxtoolsr
}

