% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{splitSifnxByPathway}
\alias{splitSifnxByPathway}
\title{Splits SIFNX files}
\usage{
splitSifnxByPathway(edges)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
an XMLInternalDocument
}
\description{
Splits SIFNX files
}
\details{
This method can be slow
}
\concept{
paxtoolsr
}

