% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{summarizeSif}
\alias{summarizeSif}
\title{Summarize a SIF Network}
\usage{
summarizeSif(sif)
}
\arguments{
\item{sif}{a binary SIF as a data.frame with three columns:
"PARTICIPANT_A", "INTERACTION_TYPE", "PARTICIPANT_B"}
}
\value{
a list containing a count of the unique genes in the SIF and counts for the interaction types in the network
}
\description{
Summarize a SIF Network
}
\examples{
results <- readSif(system.file("extdata", "test_sif.txt", package="paxtoolsr"))
summarizeSif(results)
}
\concept{
paxtoolsr
}

