% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCytoscape.R
\name{toCytoscape}
\alias{toCytoscape}
\title{Convert igraph to Cytoscape JSON}
\usage{
toCytoscape(igraphobj)
}
\arguments{
\item{igraphobj}{an igraph object}
}
\value{
a JSON object
}
\description{
Convert igraph to Cytoscape JSON
}
\note{
From https://github.com/idekerlab/cy-rest-R/blob/17f748426bb5e48ba4075b9d97318ad582b250da/utility/cytoscape_util.R
}
\examples{
library(igraph)
g <- barabasi.game(20) 
json <- toCytoscape(g)

}
\concept{
paxtoolsr
}
