% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGSEA.R
\name{toGSEA}
\alias{toGSEA}
\title{Converts a BioPAX OWL file to a GSEA GMT gene set}
\usage{
toGSEA(inputFile, outputFile = NULL, database, crossSpeciesCheckFlag)
}
\arguments{
\item{inputFile}{a string of the name of the input OWL file}

\item{outputFile}{a string of the name of the output file}

\item{database}{a string of the name of the identifier type to be included 
(e.g. "HGNC Symbol")}

\item{crossSpeciesCheckFlag}{a boolean that ensures participant protein is 
from same species}
}
\value{
see readGmt()
}
\description{
This function converts pathway information stored as BioPAX files into the 
the GSEA .gmt format.
}
\details{
The GSEA GMT format is a tab-delimited format where each row 
represents a gene set. The first column is the gene set name. The second 
column is a brief description. Other columns for each row contain genes in 
the gene set; these rows may be of unequal lengths.
}
\examples{
outFile <- tempfile()
results <- toGSEA(system.file("extdata", "biopax3-short-metabolic-pathway.owl", 
                              package="paxtoolsr"), 
                              outFile, 
                              "uniprot", 
                              crossSpeciesCheckFlag=TRUE) 

}
\concept{
paxtoolsr
}
