% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paxtools.R
\name{toSBGN}
\alias{toSBGN}
\title{Convert a BioPAX OWL file to SBGNML}
\usage{
toSBGN(inputFile, outputFile = NULL)
}
\arguments{
\item{inputFile}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string of the name of the output SBGNML file}
}
\value{
an XMLInternalDocument representing a SBGNML file
}
\description{
This function will convert a BioPAX OWL file into the Systems Biology Graphical
Notation (SBGN) Markup Language (SBGNML) XML representation
}
\details{
Objects in the SBGNML format are laid out using a Compound Spring
Embedder (CoSE) layout
}
\examples{
outFile <- tempfile()
results <- toSBGN(system.file("extdata", "biopax3-short-metabolic-pathway.owl",
  package="paxtoolsr"),
  outFile)
}
\references{
\url{http://www.cs.bilkent.edu.tr/~ivis/layout/cose-animated-demo/cose.html}
}
\concept{
paxtoolsr
}

