% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/paxtools.R
=======
% Please edit documentation in R/toSif.R
>>>>>>> devel
\name{toSif}
\alias{toSif}
\title{Convert a BioPAX OWL file to SIF}
\usage{
toSif(inputFile, outputFile = NULL)
}
\arguments{
\item{inputFile}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string of the name of the output SIF file (Optional)}
}
\value{
see readSif()
}
\description{
Convert a BioPAX OWL file to a binary SIF file
}
\details{
Information on SIF conversion is provided on the Pathway Commons
  site: \url{http://www.pathwaycommons.org/pc2/}
}
\examples{
outFile <- tempfile()
results <- toSif(system.file("extdata", "raf_map_kinase_cascade_reactome.owl",
  package="paxtoolsr"),
  outFile)
}
\concept{
paxtoolsr
}

