% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traverse.R
\name{traverse}
\alias{traverse}
\title{Access Pathway Commons using XPath-type expressions}
\usage{
traverse(uri, path, verbose = FALSE)
}
\arguments{
\item{uri}{a BioPAX element URI - specified similarly to the 'GET' command
above). Multiple IDs are allowed (uri=...&uri=...&uri=...).}

\item{path}{a BioPAX propery path in the form of
property1[:type1]/property2[:type2]; see properties, inverse properties,
Paxtools, org.biopax.paxtools.controller.PathAccessor.}

\item{verbose}{a boolean, display the command used to query Pathway Commons}
}
\value{
an XMLInternalDocument with results
}
\description{
This command provides XPath-like access to the Pathway Commons.
}
\details{
With traverse users can explicitly state the paths they would like
  to access. The format of the path query is in the form: [Initial
  Class]/[property1]:[classRestriction(optional)]/[property2]... A "*" sign
  after the property instructs path accessor to transitively traverse that
  property. For example, the following path accessor will traverse through
  all physical entity components within a complex:
  "Complex/component*/entityReference/xref:UnificationXref" The following
  will list display names of all participants of interactions, which are
  components (pathwayComponent) of a pathway (note: pathwayOrder property,
  where same or other interactions can be reached, is not considered here):
  "Pathway/pathwayComponent:Interaction/participant*/displayName" The
  optional parameter classRestriction allows to restrict/filter the returned
  property values to a certain subclass of the range of that property. In the
  first example above, this is used to get only the Unification Xrefs. Path
  accessors can use all the official BioPAX properties as well as additional
  derived classes and parameters in paxtools such as inverse parameters and
  interfaces that represent anonymous union classes in OWL. (See Paxtools
  documentation for more details).
}
\examples{
uri <- "http://identifiers.org/uniprot/P38398"
#results <- traverse(uri=uri, path="ProteinReference/organism/displayName")
}
\references{
Paxtools Documentation: \url{http://www.biopax.org/m2site/}
}
\concept{
paxtoolsr
}

