% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validate}
\alias{validate}
\title{Validate BioPAX files}
\usage{
validate(inputFile, outputFile = NULL, type = c("xml", "html", "biopax"),
  autoFix = FALSE, onlyErrors = FALSE, maxErrors = NULL,
  notStrict = FALSE)
}
\arguments{
\item{inputFile}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string of the name of the output file containing
validation results}

\item{type}{a string denoting the type of output: xml (default), html, biopax}

\item{autoFix}{a boolean that determines if the input file should be
fixed automatically. Errors that can be automatically fixed include
generating displayName properties from names, inferring organism, and
inferring dataSource}

\item{onlyErrors}{a boolean of whether to only display errors}

\item{maxErrors}{a integer denoting the number of errors to return}

\item{notStrict}{a boolean of whether to be strict in validation (default: FALSE)}
}
\value{
an XMLInternalDocument is returned if type is set to "xml" otherwise
  the location of the outputfile is returned.
}
\description{
This function validates BioPAX files for errors.
}
\details{
See the publication by Rodchenkov, et al. for information on the
  BioPAX validator. See \url{http://biopax.baderlab.org/validator} for
  additional information on validator.
  See \url{http://biopax.baderlab.org/validator/errorTypes.html} for
  information on error types.
}
\examples{
outFile <- tempfile()
rawDoc <- validate(system.file("extdata", "raf_map_kinase_cascade_reactome.owl",
  package="paxtoolsr"), onlyErrors=TRUE)
}
\references{
Rodchenkov I, Demir E, Sander C, Bader GD. The BioPAX Validator,
  \url{http://www.ncbi.nlm.nih.gov/pubmed/23918249}
}
\concept{
paxtoolsr
}

