% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDT.R
\name{convertToDT}
\alias{convertToDT}
\title{Convert Results from readSifnx to data.table}
\usage{
convertToDT(lst)
}
\arguments{
\item{lst}{a list returned from readSifnx}
}
\value{
a list entries converted to data.table
}
\description{
Convert Results from readSifnx to data.table
}
\details{
The SIFNX format is an evolving format. Older datasets may not have 
 all the columns this function expects. In these cases, the columns will be 
 added with all NULL entries.
}
\concept{
paxtoolsr
}
