% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{downloadPc}
\alias{downloadPc}
\title{Download Pathway Commons data}
\usage{
downloadPc(format = c("SIFNX", "GMT"), verbose = FALSE)
}
\arguments{
\item{format}{a string describing the format to be downloaded;
currently, only the Extended Simple Interaction Format (SIF) "SIFNX" and
Gene Set Enrichment Analysis "GMT" formats for the entire Pathway Commons
database are supported.}

\item{verbose}{a boolean debugging information}
}
\value{
a named list with named pathways, each entry contains a vector of gene
  symbols (for GMT) or a list with two data.frames (for SIFNX):
  \itemize{
    \item edges Network edges with the following columns:
      PARTICIPANT_A: Edge (interaction) participant,
      INTERACTION_TYPE: Interaction type (see details),
      PARTICIPANT_B: Edge (interaction) participant,
      INTERACTION_DATA_SOURCE: Semi-colon delimited list of database sources
        of the interaction,
      INTERACTION_PUBMED_ID: Semi-colon delimited list of NCBI Pubmed IDs
        that give evidence for the interaction
    \item nodes Node information:
      PARTICIPANT: Interaction participant,
      PARTICIPANT_TYPE: BioPAX class (see details),
      PARTICIPANT_NAME: Display name for the participant,
      UNIFICATION_XREF: A UnificationXref defines a reference to an entity
        in an external resource that has the same biological identity as
        the referring entity,
      RELATIONSHIP_XREF: An RelationshipXref defines a reference to an entity
        in an external resource that does not have the same biological
        identity as the referring entity.
  }
}
\description{
Download Pathway Commons data in various formats
}
\details{
Description of SIF interactions: http://www.pathwaycommons.org/pc/sif_interaction_rules.do
Description of BioPAX classes: http://www.biopax.org/owldoc/Level3/
}
\examples{
format <- "GMT"
#downloadPc(format=format)
}
\seealso{
\code{\link{downloadPc}}
}
\concept{
paxtoolsr
}

