% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadPc2}
\alias{downloadPc2}
\title{Download Pathway Commons files (uses menu and cache)}
\usage{
downloadPc2(destDir = NULL)
}
\arguments{
\item{destDir}{a string, the destination directory for the file to be
  downloaded (Default: NULL). If NULL, then file will be downloaded to cache
  directory file.path(Sys.getenv("HOME"), ".paxtoolsRCache")}
}
\value{
an R object using one of the read* methods provided in this package
  corresponding to the file downloaded
}
\description{
Download Pathway Commons files (uses menu and cache)
}
\examples{
\dontrun{
  downloadPc2(tempdir())
}
}
\concept{
paxtoolsr
}

