% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPc2.R
\name{downloadPc2}
\alias{downloadPc2}
\title{Download Pathway Commons files (uses menu and cache)}
\usage{
downloadPc2(selectedFileName = NULL, destDir = NULL, version = NULL,
  verbose = FALSE)
}
\arguments{
\item{selectedFileName}{a string, a name of a file to skip the the interactive selection}

\item{destDir}{a string, the destination directory for the file to be 
downloaded (Default: NULL). If NULL, then file will be downloaded to cache
directory file.path(Sys.getenv("HOME"), ".paxtoolsRCache")}

\item{version}{a version number for a previous version of Pathway Commons data; 
versions 3 and above}

\item{verbose}{a flag to display debugging information (Default: FALSE)}
}
\value{
an R object using one of the read* methods provided in this package 
  corresponding to the file downloaded
}
\description{
Download Pathway Commons files (uses menu and cache)
}
\examples{
\dontrun{
  downloadPc2(tempdir())
}
  
}
\concept{
paxtoolsr
}

