% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSignedPC.R
\name{downloadSignedPC}
\alias{downloadSignedPC}
\title{Download a SIF file containing only signed interactions}
\usage{
downloadSignedPC(destDir = NULL, forceCache = FALSE)
}
\arguments{
\item{destDir}{a string, the destination directory for the file to be 
downloaded (Default: NULL). If NULL, then file will be downloaded to cache
directory file.path(Sys.getenv("HOME"), ".paxtoolsRCache")}

\item{forceCache}{a boolean to force the use of a cached version (DEFAULT: FALSE); 
the current host of the file (GitHub) does not support the LAST-MODIFIED header}
}
\value{
a SIF containing interactions that are considered signed (i.e. 
  interactions causing an increase on decrease in a molecular species)
}
\description{
Download a SIF file containing only signed interactions
}
\examples{
# downloadSignedPC()

}
