% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paxtools.R
\name{fetch}
\alias{fetch}
\title{Fetch a set of IDs from a BioPAX OWL file}
\usage{
fetch(inputFile, outputFile = NULL, idList)
}
\arguments{
\item{inputFile}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string with the name of the output BioPAX OWL file}

\item{idList}{a vector of IDs from the BioPAX OWL file}
}
\value{
an XMLInternalDocument representing a BioPAX OWL file
}
\description{
This function will create a subsetted object with specified URIs.
}
\details{
Only entities in the input BioPAX file will be used in the fetch.
IDs used must be URIs for the entities of interest. Additional properties
such as cross-references for fetched entities will be included in the output.
}
\examples{
outFile <- tempfile()
tmp <- getPc(uri="http://identifiers.org/reactome/REACT_12034.3",
             format="BIOPAX",
             verbose=TRUE)
ids <- c("http://identifiers.org/uniprot/P36894",
         "http://identifiers.org/uniprot/Q13873")
results <- fetch(tmp, outFile, ids)
}
\concept{
paxtoolsr
}

