% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSif.R
\name{filterSif}
\alias{filterSif}
\title{Keep interactions in SIF network of certain interaction types}
\usage{
filterSif(sif, interactionTypes)
}
\arguments{
\item{sif}{a binary SIF as a data.frame with three columns: 
"PARTICIPANT_A", "INTERACTION_TYPE", "PARTICIPANT_B"}

\item{interactionTypes}{a vector of interaction types 
(List of interaction types: http://www.pathwaycommons.org/pc2/formats)}
}
\value{
a data.frame of filtered interactions with three columns: "PARTICIPANT_A", "INTERACTION_TYPE", "PARTICIPANT_B"
}
\description{
Keep interactions in SIF network of certain interaction types
}
\examples{
results <- readSif(system.file("extdata", "test_sif.txt", package="paxtoolsr"))
intTypes <- c("controls-state-change-of", "controls-expression-of", "catalysis-precedes")
filteredNetwork <- filterSif(results, intTypes)

}
\concept{
paxtoolsr
}

