% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcDatabaseNames.R
\name{getPcDatabaseNames}
\alias{getPcDatabaseNames}
\title{Get a Pathway Commons Databases}
\usage{
getPcDatabaseNames(version)
}
\arguments{
\item{version}{PC2 version}
}
\value{
a names of databases that can be used as part of queries
}
\description{
Get a Pathway Commons Databases
}
\examples{
getPcDatabaseNames(version=10)

}
\concept{
paxtoolsr
}
