% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShortestPathSif.R
\name{getShortestPathSif}
\alias{getShortestPathSif}
\title{Get the shortest between two IDs (HGNC or CHEBI)}
\usage{
getShortestPathSif(sif, idA, idB, mode = c("all", "out", "in"),
  weights = NULL, filterFun, ...)
}
\arguments{
\item{sif}{a SIF network}

\item{idA}{HGNC or CHEBI (CHEBI:XXXXX) ID}

\item{idB}{HGNC or CHEBI (CHEBI:XXXXX) ID}

\item{mode}{see shortest_paths() in igraph}

\item{weights}{see shortest_paths() in igraph}

\item{filterFun}{a function to filter multiple paths of the same length}

\item{...}{additional arguments passed on to filterFun}
}
\value{
a data.frame representing a SIF network
}
\description{
Get the shortest between two IDs (HGNC or CHEBI)
}
\concept{
paxtoolsr
}
