% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSifInIgraph.R
\name{loadSifInIgraph}
\alias{loadSifInIgraph}
\title{Load SIF as igraph Network}
\usage{
loadSifInIgraph(sif, directed = TRUE)
}
\arguments{
\item{sif}{a binary SIF as a data.frame with three columns: 
"PARTICIPANT_A", "INTERACTION_TYPE", "PARTICIPANT_B"}

\item{directed}{a boolean weather the returned graph should be directed (DEFAULT: TRUE)}
}
\value{
a directed igraph network with interaction types
}
\description{
Load SIF as igraph Network
}
\details{
Users are likely to run into issues if the input SIF has factor levels
}
\examples{
results <- readSif(system.file("extdata", "test_sif.txt", package="paxtoolsr"))
g <- loadSifInIgraph(results)

}
\concept{
paxtoolsr
}

