% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcFormats.R
\name{pcFormats}
\alias{pcFormats}
\title{Acceptable Pathway Commons Formats}
\usage{
pcFormats()
}
\value{
acceptable Pathway Commons formats
}
\description{
A simple function to see valid options
}
\details{
See references.
}
\examples{
pcFormats()

}
\references{
Output Formats Description: \url{http://www.pathwaycommons.org/pc2/help/formats.html}
}
\concept{
paxtoolsr
}

