% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathwayCommons.R
\name{processPcRequest}
\alias{processPcRequest}
\title{Process Pathway Commons request in various formats}
\usage{
processPcRequest(content, format)
}
\arguments{
\item{content}{a string, content to be processed}

\item{format}{a string, the type of format}
}
\description{
Process Pathway Commons request in various formats
}
\examples{
fileName <- system.file("extdata", "test_biopax.owl", package="paxtoolsr")
content <- readChar(fileName, file.info(fileName)$size)
results <- processPcRequest(content, "BIOPAX")
}
\seealso{
\code{\link{pcFormats}}
}
\concept{
paxtoolsr
}

