% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiopax.R
\name{readBiopax}
\alias{readBiopax}
\title{Read BioPAX files as XML documents}
\usage{
readBiopax(inputFile)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
an XMLInternalDocument
}
\description{
Read BioPAX files as XML documents
}
\examples{
results <- readBiopax(system.file("extdata", "biopax3-short-metabolic-pathway.owl", 
  package="paxtoolsr"))

}
\concept{
paxtoolsr
}

