% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/paxtools.R
=======
% Please edit documentation in R/readGmt.R
>>>>>>> devel
\name{readGmt}
\alias{readGmt}
\title{Read in gene sets from GMT files}
\usage{
readGmt(inputFile)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
a named list where each entry corresponds to a gene set
}
\description{
This function will read in gene sets in the GMT format into a named list.
}
\examples{
results <- readGmt(system.file("extdata", "test_gsea.gmt", package="paxtoolsr"))
}
\concept{
paxtoolsr
}

