% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{readSbgn}
\alias{readSbgn}
\title{Read SBGN files as XML documents}
\usage{
readSbgn(inputFile)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
an XMLInternalDocument
}
\description{
Read SBGN files as XML documents
}
\examples{
results <- readSbgn(system.file("extdata", "test_sbgn.xml", package="paxtoolsr"))
}
\concept{
paxtoolsr
}

