% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSifnx.R
\name{readSifnx}
\alias{readSifnx}
\title{Read in a Extended SIF file}
\usage{
readSifnx(inputFile)
}
\arguments{
\item{inputFile}{an inputFile}
}
\value{
a list with nodes and edges entries
}
\description{
Read in a Extended SIF file
}
\details{
SIFNX files from Pathway Commons commonly come a single file that
includes a tab-delimited sections for nodes and another for edges. The
sections are separated by an empty lines. These sections must be split before
they are read.
}
\examples{
results <- readSifnx(system.file("extdata", "test_sifnx.txt", package="paxtoolsr"))

}
\concept{
paxtoolsr
}
