% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchListOfVectors.R
\name{searchListOfVectors}
\alias{searchListOfVectors}
\title{Search List of Vectors}
\usage{
searchListOfVectors(q, lst)
}
\arguments{
\item{q}{query vector}

\item{lst}{list of vectors to search}
}
\value{
a list of vectors with the same length as the query vector, each list
  entry will have indicies for lst where there was a match with the query
  vector. Return NA if there were no matches.
}
\description{
Search List of Vectors
}
\details{
Taken from: http://stackoverflow.com/questions/11002391/fast-way-of-getting-index-of-match-in-list
}
\examples{
lst <- list(1:3, 3:5, 3:7)
q <- c(3, 5)
results <- searchListOfVectors(q, lst)
names(results) <- q
}
\concept{
paxtoolsr
}

