% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{searchListOfVectors}
\alias{searchListOfVectors}
\title{Search List of Vectors}
\usage{
searchListOfVectors(q, lst)
}
\arguments{
\item{q}{query vector}

\item{lst}{list of vectors to search}
}
\value{
a list of vectors with the same length as the query vector, each list
  entry will have indicies for lst where there was a match with the query
  vector
}
\description{
Search List of Vectors
}
\examples{
lst <- list(1:3, 3:5, 3:7)
q <- c(3, 5)
results <- searchListOfVectors(q, lst)
}
\concept{
paxtoolsr
}

