% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toLevel3.R
\name{toLevel3}
\alias{toLevel3}
\title{Convert a BioPAX OWL file to BioPAX Level 3}
\usage{
toLevel3(inputFile, outputFile = NULL)
}
\arguments{
\item{inputFile}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string of the name of the output BioPAX OWL file}
}
\value{
an XMLInternalDocument representing a BioPAX OWL file
}
\description{
This file will convert older BioPAX objects to BioPAX Level 3
}
\examples{
outFile <- tempfile()
results <- toLevel3(system.file("extdata", "raf_map_kinase_cascade_reactome.owl",
  package="paxtoolsr"),
  outFile)
}
\concept{
paxtoolsr
}

