% Generated by roxygen2 (4.1.1): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/paxtools.R
=======
% Please edit documentation in R/toLevel3.R
>>>>>>> devel
\name{toLevel3}
\alias{toLevel3}
\title{Convert a PSIMI or older BioPAX OWL file to BioPAX Level 3}
\usage{
toLevel3(inputFile, outputFile = NULL)
}
\arguments{
\item{inputFile}{a string of the name of the input file}

\item{outputFile}{a string of the name of the output BioPAX OWL file}
}
\value{
an XMLInternalDocument representing a BioPAX OWL file
}
\description{
This file will convert PSIMI or older BioPAX objects to BioPAX Level 3
}
\examples{
inputFile <- system.file("extdata", "raf_map_kinase_cascade_reactome.owl",
  package="paxtoolsr")
outFile <- tempfile()
<<<<<<< HEAD
results <- toLevel3(system.file("extdata", "raf_map_kinase_cascade_reactome.owl",
  package="paxtoolsr"),
  outFile)
=======
results <- toLevel3(inputFile, outFile)
>>>>>>> devel
}
\concept{
paxtoolsr
}

