% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topPathways.R
\name{topPathways}
\alias{topPathways}
\title{Retrieve top pathways}
\usage{
topPathways(q = NULL, datasource = NULL, organism = NULL,
  verbose = FALSE)
}
\arguments{
\item{q}{[Optional] a keyword, name, external identifier, or a Lucene query string, like in 'search', but the default is '*' (match all).}

\item{datasource}{filter by data source (same as for 'search').}

\item{organism}{organism filter (same as for 'search').}

\item{verbose}{a boolean, display the command used to query Pathway Commons}
}
\value{
a data.frame with the following columns:
  \itemize{ 
    \item uri URI ID for the pathway
    \item biopaxClass the type of BioPAX object
    \item name a human readable name
    \item dataSource the dataSource for the pathway 
    \item organism an organism identifier 
    \item pathway URI ID for the pathway
  }
}
\description{
This command returns all "top" pathways.
}
\details{
Pathways that are neither 'controlled' nor 'pathwayComponent' of 
  another process.
}
\examples{
#results <- topPathways(datasource="panther")

}
\concept{
paxtoolsr
}
