% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toSif}
\alias{toSif}
\title{Convert a BioPAX OWL file to SIF}
\usage{
toSif(inputFile, outputFile = NULL)
}
\arguments{
\item{inputFile}{a string of the name of the input BioPAX OWL file}

\item{outputFile}{a string of the name of the output SIF file (Optional)}
}
\value{
a 3-column data.frame where the columns are named: PARTICIPANT_A,
  INTERACTION_TYPE, PARTICIPANT_B
}
\description{
Convert a BioPAX OWL file to a binary SIF file
}
\details{
Information on SIF conversion is provided on the Pathway Commons
  site: \url{http://www.pathwaycommons.org/pc2/}
}
\examples{
outFile <- tempfile()
results <- toSif(system.file("extdata", "dna_replication.owl",
  package="paxtoolsr"),
  outFile)
}
\concept{
paxtoolsr
}

