% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/toSifnx.R
\name{toSifnx}
\alias{toSifnx}
\title{Converts BioPAX OWL file to extended binary SIF representation}
\usage{
toSifnx(inputFile, outputFile = tempfile())
}
\arguments{
\item{inputFile}{a string with the name of the input BioPAX OWL file}

\item{outputFile}{a string with the name of the output file for SIFNX
  information}
}
\value{
see readSifnx()
}
\description{
Converts BioPAX OWL file to extended binary SIF representation
}
\details{
Information on SIF conversion is provided on the Pathway Commons
  site: \url{http://www.pathwaycommons.org/pc2/}
}
\examples{
inputFile <- system.file("extdata", "raf_map_kinase_cascade_reactome.owl", package="paxtoolsr")
results <- toSifnx(inputFile=inputFile)
}
\concept{
paxtoolsr
}

